icotool: Use size_t when counting array elements, not int.
It's not safe to use int when counting array elements, as on 64 bit
platforms the arrays can overflow a 32 bit int.  (That is, in general.
In this particular case it probably is safe because the original array
is derived from argv.  But it's easier to use the correct type.)

+++ icotool/main.c
@@ -217,7 +217,7 @@ main(int argc, char **argv)
     bool create_mode = false;
     FILE *in;
     char *inname;
-    int raw_filec = 0;
+    size_t raw_filec = 0;
     char** raw_filev = 0;
 
     set_program_name(argv[0]);
